backend/requirements.txt
Django==3.2.4
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.12.2
django-redis==4.12.1
celery==4.1.1
drf-yasg==1.20.0
psycopg2-binary==2.8.5
django-filter==2.4.0
Faker==8.10.1
factory-boy==3.2.0
pytest-mock==3.6.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
